Move the arguments module into cli/
