+++ lib/ansible/cli/arguments/optparse_helpers.py
